reverts and fixes
