+++ maze_dataset/token_utils.py
@@ -10,6 +10,7 @@ import numpy as np
 from jaxtyping import Bool, Float, Int, Int8
 from muutils.errormode import ErrorMode
 from muutils.misc import list_join
+from muutils.misc.sequence import WhenMissing
 
 from maze_dataset.constants import (
     CARDINAL_MAP,
@@ -19,7 +20,6 @@ from maze_dataset.constants import (
     ConnectionList,
     CoordTup,
 )
-from maze_dataset.utils import WhenMissing
 
 # filtering things from a prompt or generated text
 # ==================================================
